refactor(api)!: Refactor the vulnerability resolution component
Align all endpoints for vulnerability resolutions with the rest of the
back-end codebase:
- rename operation IDs
- rename request objects to comply with the operation IDs
- change test output to comply with the new operation IDs
- rename files accordingly
- align the query client usage in the UI to comply with the changed
  function names due to OpenAPI changes

Resolves #4658.
