+++ components/resolutions/api-model/src/commonMain/kotlin/PatchVulnerabilityResolution.kt
@@ -24,7 +24,7 @@ import kotlinx.serialization.Serializable
 import org.eclipse.apoapsis.ortserver.shared.apimodel.VulnerabilityResolutionReason
 
 @Serializable
-data class CreateVulnerabilityResolution(
-    val reason: VulnerabilityResolutionReason,
-    val comment: String
+data class PatchVulnerabilityResolution(
+    val reason: VulnerabilityResolutionReason? = null,
+    val comment: String? = null
 )
