+++ components/resolutions/api-model/src/commonMain/kotlin/PostVulnerabilityResolution.kt
@@ -24,7 +24,7 @@ import kotlinx.serialization.Serializable
 import org.eclipse.apoapsis.ortserver.shared.apimodel.VulnerabilityResolutionReason
 
 @Serializable
-data class UpdateVulnerabilityResolution(
-    val reason: VulnerabilityResolutionReason? = null,
-    val comment: String? = null
+data class PostVulnerabilityResolution(
+    val reason: VulnerabilityResolutionReason,
+    val comment: String
 )
