+++ components/resolutions/backend/src/routes/kotlin/routes/vulnerabilities/PatchVulnerabilityResolution.kt
@@ -29,9 +29,9 @@ import io.ktor.server.routing.Route
 
 import org.eclipse.apoapsis.ortserver.components.authorization.rights.RepositoryPermission
 import org.eclipse.apoapsis.ortserver.components.authorization.routes.OrtServerPrincipal.Companion.requirePrincipal
-import org.eclipse.apoapsis.ortserver.components.authorization.routes.post
+import org.eclipse.apoapsis.ortserver.components.authorization.routes.patch
 import org.eclipse.apoapsis.ortserver.components.authorization.routes.requirePermission
-import org.eclipse.apoapsis.ortserver.components.resolutions.CreateVulnerabilityResolution
+import org.eclipse.apoapsis.ortserver.components.resolutions.PatchVulnerabilityResolution
 import org.eclipse.apoapsis.ortserver.components.resolutions.vulnerabilities.VulnerabilityResolutionError
 import org.eclipse.apoapsis.ortserver.components.resolutions.vulnerabilities.VulnerabilityResolutionService
 import org.eclipse.apoapsis.ortserver.model.RepositoryId
@@ -41,14 +41,14 @@ import org.eclipse.apoapsis.ortserver.shared.ktorutils.jsonBody
 import org.eclipse.apoapsis.ortserver.shared.ktorutils.requireIdParameter
 import org.eclipse.apoapsis.ortserver.shared.ktorutils.requireParameter
 
-internal fun Route.createVulnerabilityResolution(
+internal fun Route.patchVulnerabilityResolution(
     vulnerabilityResolutionService: VulnerabilityResolutionService
-) = post(
+) = patch(
     "repositories/{repositoryId}/resolutions/vulnerabilities/{externalId}",
     {
-        operationId = "CreateVulnerabilityResolution"
-        summary = "Create a vulnerability resolution for a vulnerability in a repository."
-        description = "Create a vulnerability resolution for a vulnerability in a repository."
+        operationId = "PatchVulnerabilityResolution"
+        summary = "Update a vulnerability resolution for a vulnerability in a repository."
+        description = "Update a vulnerability resolution for a vulnerability in a repository."
         tags = listOf("Resolutions")
 
         request {
@@ -62,12 +62,12 @@ internal fun Route.createVulnerabilityResolution(
                 required = true
             }
 
-            jsonBody<CreateVulnerabilityResolution> {
-                description = "The details of the vulnerability resolution to create."
+            jsonBody<PatchVulnerabilityResolution> {
+                description = "The details of the vulnerability resolution to update."
                 required = true
 
                 example("Example") {
-                    value = CreateVulnerabilityResolution(
+                    value = PatchVulnerabilityResolution(
                         reason = VulnerabilityResolutionReason.INEFFECTIVE_VULNERABILITY,
                         comment = "This vulnerability is not relevant because it affects code that is not used."
                     )
@@ -76,8 +76,8 @@ internal fun Route.createVulnerabilityResolution(
         }
 
         response {
-            HttpStatusCode.Created to {
-                description = "The vulnerability resolution was successfully created."
+            HttpStatusCode.OK to {
+                description = "The vulnerability resolution was updated successfully."
             }
 
             HttpStatusCode.BadRequest to {
@@ -85,7 +85,7 @@ internal fun Route.createVulnerabilityResolution(
             }
 
             HttpStatusCode.NotFound to {
-                description = "The repository or vulnerability was not found."
+                description = "The repository or vulnerability resolution was not found."
             }
         }
     },
@@ -95,16 +95,16 @@ internal fun Route.createVulnerabilityResolution(
     val repositoryId = call.requireIdParameter("repositoryId")
     val externalId = call.requireParameter("externalId")
 
-    val createVulnerabilityResolution = call.receive<CreateVulnerabilityResolution>()
+    val patchVulnerabilityResolution = call.receive<PatchVulnerabilityResolution>()
 
-    vulnerabilityResolutionService.createResolution(
+    vulnerabilityResolutionService.updateResolution(
         repositoryId = RepositoryId(repositoryId),
         externalId = externalId,
         userId = userId,
-        reason = createVulnerabilityResolution.reason.mapToModel(),
-        comment = createVulnerabilityResolution.comment
+        reason = patchVulnerabilityResolution.reason?.mapToModel(),
+        comment = patchVulnerabilityResolution.comment
     ).onOk {
-        call.respond(HttpStatusCode.Created)
+        call.respond(HttpStatusCode.OK)
     }.onErr {
         when (it) {
             is VulnerabilityResolutionError.InvalidState -> call.respond(HttpStatusCode.BadRequest, it.message)
