+++ components/resolutions/backend/src/routes/kotlin/routes/vulnerabilities/PostVulnerabilityResolution.kt
@@ -29,9 +29,9 @@ import io.ktor.server.routing.Route
 
 import org.eclipse.apoapsis.ortserver.components.authorization.rights.RepositoryPermission
 import org.eclipse.apoapsis.ortserver.components.authorization.routes.OrtServerPrincipal.Companion.requirePrincipal
-import org.eclipse.apoapsis.ortserver.components.authorization.routes.patch
+import org.eclipse.apoapsis.ortserver.components.authorization.routes.post
 import org.eclipse.apoapsis.ortserver.components.authorization.routes.requirePermission
-import org.eclipse.apoapsis.ortserver.components.resolutions.UpdateVulnerabilityResolution
+import org.eclipse.apoapsis.ortserver.components.resolutions.PostVulnerabilityResolution
 import org.eclipse.apoapsis.ortserver.components.resolutions.vulnerabilities.VulnerabilityResolutionError
 import org.eclipse.apoapsis.ortserver.components.resolutions.vulnerabilities.VulnerabilityResolutionService
 import org.eclipse.apoapsis.ortserver.model.RepositoryId
@@ -41,14 +41,14 @@ import org.eclipse.apoapsis.ortserver.shared.ktorutils.jsonBody
 import org.eclipse.apoapsis.ortserver.shared.ktorutils.requireIdParameter
 import org.eclipse.apoapsis.ortserver.shared.ktorutils.requireParameter
 
-internal fun Route.updateVulnerabilityResolution(
+internal fun Route.postVulnerabilityResolution(
     vulnerabilityResolutionService: VulnerabilityResolutionService
-) = patch(
+) = post(
     "repositories/{repositoryId}/resolutions/vulnerabilities/{externalId}",
     {
-        operationId = "UpdateVulnerabilityResolution"
-        summary = "Update a vulnerability resolution for a vulnerability in a repository."
-        description = "Update a vulnerability resolution for a vulnerability in a repository."
+        operationId = "PostVulnerabilityResolution"
+        summary = "Create a vulnerability resolution for a vulnerability in a repository."
+        description = "Create a vulnerability resolution for a vulnerability in a repository."
         tags = listOf("Resolutions")
 
         request {
@@ -62,12 +62,12 @@ internal fun Route.updateVulnerabilityResolution(
                 required = true
             }
 
-            jsonBody<UpdateVulnerabilityResolution> {
-                description = "The details of the vulnerability resolution to update."
+            jsonBody<PostVulnerabilityResolution> {
+                description = "The details of the vulnerability resolution to create."
                 required = true
 
                 example("Example") {
-                    value = UpdateVulnerabilityResolution(
+                    value = PostVulnerabilityResolution(
                         reason = VulnerabilityResolutionReason.INEFFECTIVE_VULNERABILITY,
                         comment = "This vulnerability is not relevant because it affects code that is not used."
                     )
@@ -76,8 +76,8 @@ internal fun Route.updateVulnerabilityResolution(
         }
 
         response {
-            HttpStatusCode.OK to {
-                description = "The vulnerability resolution was updated successfully."
+            HttpStatusCode.Created to {
+                description = "The vulnerability resolution was successfully created."
             }
 
             HttpStatusCode.BadRequest to {
@@ -85,7 +85,7 @@ internal fun Route.updateVulnerabilityResolution(
             }
 
             HttpStatusCode.NotFound to {
-                description = "The repository or vulnerability resolution was not found."
+                description = "The repository or vulnerability was not found."
             }
         }
     },
@@ -95,16 +95,16 @@ internal fun Route.updateVulnerabilityResolution(
     val repositoryId = call.requireIdParameter("repositoryId")
     val externalId = call.requireParameter("externalId")
 
-    val updateVulnerabilityResolution = call.receive<UpdateVulnerabilityResolution>()
+    val postVulnerabilityResolution = call.receive<PostVulnerabilityResolution>()
 
-    vulnerabilityResolutionService.updateResolution(
+    vulnerabilityResolutionService.createResolution(
         repositoryId = RepositoryId(repositoryId),
         externalId = externalId,
         userId = userId,
-        reason = updateVulnerabilityResolution.reason?.mapToModel(),
-        comment = updateVulnerabilityResolution.comment
+        reason = postVulnerabilityResolution.reason.mapToModel(),
+        comment = postVulnerabilityResolution.comment
     ).onOk {
-        call.respond(HttpStatusCode.OK)
+        call.respond(HttpStatusCode.Created)
     }.onErr {
         when (it) {
             is VulnerabilityResolutionError.InvalidState -> call.respond(HttpStatusCode.BadRequest, it.message)
