+++ components/resolutions/backend/src/test/kotlin/routes/vulnerabilities/PatchVulnerabilityResolutionIntegrationTest.kt
@@ -23,11 +23,12 @@ import io.kotest.assertions.ktor.client.shouldHaveStatus
 import io.kotest.matchers.collections.containExactly
 import io.kotest.matchers.should
 
-import io.ktor.client.request.post
+import io.ktor.client.request.patch
 import io.ktor.client.request.setBody
 import io.ktor.http.HttpStatusCode
 
-import org.eclipse.apoapsis.ortserver.components.resolutions.CreateVulnerabilityResolution
+import org.eclipse.apoapsis.ortserver.components.resolutions.PatchVulnerabilityResolution
+import org.eclipse.apoapsis.ortserver.components.resolutions.PostVulnerabilityResolution
 import org.eclipse.apoapsis.ortserver.components.resolutions.beOk
 import org.eclipse.apoapsis.ortserver.components.resolutions.routes.ResolutionsIntegrationTest
 import org.eclipse.apoapsis.ortserver.components.resolutions.shouldBeOk
@@ -37,58 +38,15 @@ import org.eclipse.apoapsis.ortserver.model.runs.repository.VulnerabilityResolut
 import org.eclipse.apoapsis.ortserver.model.runs.repository.VulnerabilityResolutionReason
 import org.eclipse.apoapsis.ortserver.shared.apimodel.VulnerabilityResolutionReason as ApiVulnerabilityResolutionReason
 
-class CreateVulnerabilityResolutionIntegrationTest : ResolutionsIntegrationTest({
+class PatchVulnerabilityResolutionIntegrationTest : ResolutionsIntegrationTest({
     var repositoryId = RepositoryId(-1)
 
     beforeEach {
         repositoryId = RepositoryId(dbExtension.fixtures.repository.id)
     }
 
-    "CreateVulnerabilityResolution" should {
-        "create a vulnerability resolution for a repository" {
-            resolutionsTestApplication { client ->
-                val response = client.post(
-                    "/repositories/${repositoryId.value}/resolutions/vulnerabilities/CVE-2021-1234"
-                ) {
-                    setBody(
-                        CreateVulnerabilityResolution(
-                            comment = "This is not a vulnerability.",
-                            reason = ApiVulnerabilityResolutionReason.NOT_A_VULNERABILITY
-                        )
-                    )
-                }
-
-                response shouldHaveStatus HttpStatusCode.Created
-
-                vulnerabilityResolutionService.getResolutionsForRepository(repositoryId).shouldBeOk {
-                    it should containExactly(
-                        VulnerabilityResolution(
-                            externalId = "CVE-2021-1234",
-                            comment = "This is not a vulnerability.",
-                            reason = VulnerabilityResolutionReason.NOT_A_VULNERABILITY,
-                            source = ResolutionSource.SERVER
-                        )
-                    )
-                }
-            }
-        }
-
-        "fail if the repository does not exist" {
-            resolutionsTestApplication { client ->
-                val response = client.post("/repositories/9999/resolutions/vulnerabilities/CVE-2021-1234") {
-                    setBody(
-                        CreateVulnerabilityResolution(
-                            comment = "This is not a vulnerability.",
-                            reason = ApiVulnerabilityResolutionReason.NOT_A_VULNERABILITY
-                        )
-                    )
-                }
-
-                response shouldHaveStatus HttpStatusCode.NotFound
-            }
-        }
-
-        "fail if there is already a resolution for the vulnerability" {
+    "PatchVulnerabilityResolution" should {
+        "update a vulnerability resolution for a repository" {
             resolutionsTestApplication { client ->
                 vulnerabilityResolutionService.createResolution(
                     repositoryId = repositoryId,
@@ -98,18 +56,61 @@ class CreateVulnerabilityResolutionIntegrationTest : ResolutionsIntegrationTest(
                     userId = "test-user"
                 ) should beOk()
 
-                val response = client.post(
+                val response = client.patch(
                     "/repositories/${repositoryId.value}/resolutions/vulnerabilities/CVE-2021-1234"
                 ) {
                     setBody(
-                        CreateVulnerabilityResolution(
-                            comment = "This is not a vulnerability.",
-                            reason = ApiVulnerabilityResolutionReason.NOT_A_VULNERABILITY
+                        PatchVulnerabilityResolution(
+                            comment = "This vulnerability does not apply to our project.",
+                            reason = ApiVulnerabilityResolutionReason.INEFFECTIVE_VULNERABILITY
                         )
                     )
                 }
 
-                response shouldHaveStatus HttpStatusCode.BadRequest
+                response shouldHaveStatus HttpStatusCode.OK
+
+                vulnerabilityResolutionService.getResolutionsForRepository(repositoryId).shouldBeOk {
+                    it should containExactly(
+                        VulnerabilityResolution(
+                            externalId = "CVE-2021-1234",
+                            comment = "This vulnerability does not apply to our project.",
+                            reason = VulnerabilityResolutionReason.INEFFECTIVE_VULNERABILITY,
+                            source = ResolutionSource.SERVER
+                        )
+                    )
+                }
+            }
+        }
+
+        "fail if the repository does not exist" {
+            resolutionsTestApplication { client ->
+                val response = client.patch("/repositories/9999/resolutions/vulnerabilities/CVE-2021-1234") {
+                    setBody(
+                        PostVulnerabilityResolution(
+                            comment = "This vulnerability does not apply to our project.",
+                            reason = ApiVulnerabilityResolutionReason.INEFFECTIVE_VULNERABILITY
+                        )
+                    )
+                }
+
+                response shouldHaveStatus HttpStatusCode.NotFound
+            }
+        }
+
+        "fail if the vulnerability resolution does not exist" {
+            resolutionsTestApplication { client ->
+                val response = client.patch(
+                    "/repositories/${repositoryId.value}/resolutions/vulnerabilities/CVE-2021-9999"
+                ) {
+                    setBody(
+                        PostVulnerabilityResolution(
+                            comment = "This vulnerability does not apply to our project.",
+                            reason = ApiVulnerabilityResolutionReason.INEFFECTIVE_VULNERABILITY
+                        )
+                    )
+                }
+
+                response shouldHaveStatus HttpStatusCode.NotFound
             }
         }
     }
