+++ components/resolutions/backend/src/test/kotlin/routes/vulnerabilities/PostVulnerabilityResolutionIntegrationTest.kt
@@ -23,12 +23,11 @@ import io.kotest.assertions.ktor.client.shouldHaveStatus
 import io.kotest.matchers.collections.containExactly
 import io.kotest.matchers.should
 
-import io.ktor.client.request.patch
+import io.ktor.client.request.post
 import io.ktor.client.request.setBody
 import io.ktor.http.HttpStatusCode
 
-import org.eclipse.apoapsis.ortserver.components.resolutions.CreateVulnerabilityResolution
-import org.eclipse.apoapsis.ortserver.components.resolutions.UpdateVulnerabilityResolution
+import org.eclipse.apoapsis.ortserver.components.resolutions.PostVulnerabilityResolution
 import org.eclipse.apoapsis.ortserver.components.resolutions.beOk
 import org.eclipse.apoapsis.ortserver.components.resolutions.routes.ResolutionsIntegrationTest
 import org.eclipse.apoapsis.ortserver.components.resolutions.shouldBeOk
@@ -38,43 +37,35 @@ import org.eclipse.apoapsis.ortserver.model.runs.repository.VulnerabilityResolut
 import org.eclipse.apoapsis.ortserver.model.runs.repository.VulnerabilityResolutionReason
 import org.eclipse.apoapsis.ortserver.shared.apimodel.VulnerabilityResolutionReason as ApiVulnerabilityResolutionReason
 
-class UpdateVulnerabilityResolutionIntegrationTest : ResolutionsIntegrationTest({
+class PostVulnerabilityResolutionIntegrationTest : ResolutionsIntegrationTest({
     var repositoryId = RepositoryId(-1)
 
     beforeEach {
         repositoryId = RepositoryId(dbExtension.fixtures.repository.id)
     }
 
-    "UpdateVulnerabilityResolution" should {
-        "update a vulnerability resolution for a repository" {
+    "PostVulnerabilityResolution" should {
+        "create a vulnerability resolution for a repository" {
             resolutionsTestApplication { client ->
-                vulnerabilityResolutionService.createResolution(
-                    repositoryId = repositoryId,
-                    externalId = "CVE-2021-1234",
-                    reason = VulnerabilityResolutionReason.NOT_A_VULNERABILITY,
-                    comment = "This is not a vulnerability.",
-                    userId = "test-user"
-                ) should beOk()
-
-                val response = client.patch(
+                val response = client.post(
                     "/repositories/${repositoryId.value}/resolutions/vulnerabilities/CVE-2021-1234"
                 ) {
                     setBody(
-                        UpdateVulnerabilityResolution(
-                            comment = "This vulnerability does not apply to our project.",
-                            reason = ApiVulnerabilityResolutionReason.INEFFECTIVE_VULNERABILITY
+                        PostVulnerabilityResolution(
+                            comment = "This is not a vulnerability.",
+                            reason = ApiVulnerabilityResolutionReason.NOT_A_VULNERABILITY
                         )
                     )
                 }
 
-                response shouldHaveStatus HttpStatusCode.OK
+                response shouldHaveStatus HttpStatusCode.Created
 
                 vulnerabilityResolutionService.getResolutionsForRepository(repositoryId).shouldBeOk {
                     it should containExactly(
                         VulnerabilityResolution(
                             externalId = "CVE-2021-1234",
-                            comment = "This vulnerability does not apply to our project.",
-                            reason = VulnerabilityResolutionReason.INEFFECTIVE_VULNERABILITY,
+                            comment = "This is not a vulnerability.",
+                            reason = VulnerabilityResolutionReason.NOT_A_VULNERABILITY,
                             source = ResolutionSource.SERVER
                         )
                     )
@@ -84,11 +75,11 @@ class UpdateVulnerabilityResolutionIntegrationTest : ResolutionsIntegrationTest(
 
         "fail if the repository does not exist" {
             resolutionsTestApplication { client ->
-                val response = client.patch("/repositories/9999/resolutions/vulnerabilities/CVE-2021-1234") {
+                val response = client.post("/repositories/9999/resolutions/vulnerabilities/CVE-2021-1234") {
                     setBody(
-                        CreateVulnerabilityResolution(
-                            comment = "This vulnerability does not apply to our project.",
-                            reason = ApiVulnerabilityResolutionReason.INEFFECTIVE_VULNERABILITY
+                        PostVulnerabilityResolution(
+                            comment = "This is not a vulnerability.",
+                            reason = ApiVulnerabilityResolutionReason.NOT_A_VULNERABILITY
                         )
                     )
                 }
@@ -97,20 +88,28 @@ class UpdateVulnerabilityResolutionIntegrationTest : ResolutionsIntegrationTest(
             }
         }
 
-        "fail if the vulnerability resolution does not exist" {
+        "fail if there is already a resolution for the vulnerability" {
             resolutionsTestApplication { client ->
-                val response = client.patch(
-                    "/repositories/${repositoryId.value}/resolutions/vulnerabilities/CVE-2021-9999"
+                vulnerabilityResolutionService.createResolution(
+                    repositoryId = repositoryId,
+                    externalId = "CVE-2021-1234",
+                    reason = VulnerabilityResolutionReason.NOT_A_VULNERABILITY,
+                    comment = "This is not a vulnerability.",
+                    userId = "test-user"
+                ) should beOk()
+
+                val response = client.post(
+                    "/repositories/${repositoryId.value}/resolutions/vulnerabilities/CVE-2021-1234"
                 ) {
                     setBody(
-                        CreateVulnerabilityResolution(
-                            comment = "This vulnerability does not apply to our project.",
-                            reason = ApiVulnerabilityResolutionReason.INEFFECTIVE_VULNERABILITY
+                        PostVulnerabilityResolution(
+                            comment = "This is not a vulnerability.",
+                            reason = ApiVulnerabilityResolutionReason.NOT_A_VULNERABILITY
                         )
                     )
                 }
 
-                response shouldHaveStatus HttpStatusCode.NotFound
+                response shouldHaveStatus HttpStatusCode.BadRequest
             }
         }
     }
